Updated jetstream messaging client to use publish instead of publishAsync
After testing for a longer period of time, there would be messages "stalled with too many outstanding async published messages" which is caused by too many messages waiting to be sent. This should handle this queue by timingout in the current context instead of blocking all further publishes. This should not affect processing other events as they run in a pool reguardless.